magent changes - sort into directories
